refactor: stockTrading 타입 분리에 의한 타입 경로 수정
