+++ Client/src/components/StockInfo/styled.ts
@@ -3,12 +3,8 @@ import { StyledCommonflexCenter } from '../../style/common.styled';
 import { ReactComponent as DeleteBtnImg } from '../../assets/icons/btn-close-gray.svg';
 import { ReactComponent as UpArrowIconImg } from '../../assets/icons/icon_up.svg';
 import { ReactComponent as DownArrowIconImg } from '../../assets/icons/icon_down.svg';
-import {
-  CellColorProps,
-  StockDetailTabButtonProps,
-  StockPriceProps,
-} from '../../types/stock';
-
+import { StockDetailTabButtonProps, StockPriceProps } from '../../types/stock';
+import { CellColorProps } from '../../types/stockTrading';
 export const StockTableContainer = styled.div`
   width: 100%;
